requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

awscli==1.27.148
boto3==1.26.148
botocore==1.29.148
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.4
docutils==0.16
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gspread==5.9.0
gspread-dataframe==3.3.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.0.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16